week 2 start
finished fire attacks works with Dot, added slow attacks that work, started working on collectables ** should work need object and values

+++ First Person Experience/Assets/Scripts/Health.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    public int hp;
-    int MaxHp;
+    public float hp;
+    float MaxHp;
     public bool respawn;
     public float respawnTime;
     bool active;
